requirements.txt
Django>=4.2.0
django-environ>=0.11.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
django-redis>=5.0.0
djangorestframework-simplejwt>=5.0.0
psycopg2-binary>=2.9.0
